drop 802.1q tagged frames on LBR sessions
A LBR sessions started on a primary interface should drop
and not reply to tagged frames, as session was intented for
a VLAN.

Because tags are dropped by default by the network stack
even on raw sockets, we need to check this through packet
auxdata.

We have to decide what to do if we get tags on LBR sessions
with only PCP value and no VLAN ID, as right now they are also dropped.

Signed-off-by: Beniamin Sandu <beniaminsandu@gmail.com>