library: implement a proper fix for the timer cleanup
The problem was a possible race condition between a running interval
timer callback thread and the session cleanup handler, which will
delete the timer. The initial workaround was using a dirty sleep
and hoping for the best.

Since the timeout handler is never firing overlaping threads because
of our session logic, make sure to make the callback threads joinable,
and just join the last (possible) running one, before removing the timer,
in the cleanup handler.

Signed-off-by: Beniamin Sandu <beniaminsandu@gmail.com>